@@ -1,6 +1,6 @@
 {
     "name": "@casual-simulation/causal-tree-cli",
-    "version": "1.5.11-alpha.43",
+    "version": "2.0.0",
     "description": "A CLI to manage causal trees",
     "keywords": [
         "cli",
@@ -38,10 +38,10 @@
         "access": "public"
     },
     "dependencies": {
-        "@casual-simulation/causal-tree-client-apiary": "^1.5.11-alpha.43",
-        "@casual-simulation/causal-tree-store-cassandradb": "^1.5.11-alpha.43",
-        "@casual-simulation/causal-tree-store-mongodb": "^1.5.11-alpha.43",
-        "@casual-simulation/causal-trees": "^1.5.11-alpha.43",
+        "@casual-simulation/causal-tree-client-apiary": "^2.0.0",
+        "@casual-simulation/causal-tree-store-cassandradb": "^2.0.0",
+        "@casual-simulation/causal-tree-store-mongodb": "^2.0.0",
+        "@casual-simulation/causal-trees": "^2.0.0",
         "cassandra-driver": "4.5.1",
         "cli-progress": "3.8.1",
         "colors": "1.4.0",
@@ -52,9 +52,9 @@
     },
     "devDependencies": {
         "typescript": "^4.2.2",
-        "webpack": "^5.10.0",
-        "webpack-cli": "^4.2.0",
-        "webpack-merge": "^5.4.0",
+        "webpack": "^5.38.1",
+        "webpack-cli": "^4.7.2",
+        "webpack-merge": "^5.8.0",
         "webpack-node-externals": "^2.5.2"
     }
 }
