@@ -11,9 +11,10 @@ import {
     DEFAULT_PORTAL_ZOOMABLE,
     DEFAULT_PORTAL_ROTATABLE,
     DEFAULT_PORTAL_PANNABLE,
-    DEFAULT_INVENTORY_PORTAL_RESIZABLE,
-    DEFAULT_INVENTORY_PORTAL_HEIGHT,
+    DEFAULT_MINI_PORTAL_RESIZABLE,
+    DEFAULT_MINI_PORTAL_HEIGHT,
     Bot,
+    MINI_PORTAL,
 } from '@casual-simulation/aux-common';
 import { Simulation3D } from '../../shared/scene/Simulation3D';
 import {
@@ -22,162 +23,169 @@ import {
     watchPortalConfigBot,
 } from '@casual-simulation/aux-vm-browser';
 import { tap, filter } from 'rxjs/operators';
-import {
-    InventoryContextGroup3D as InventoryDimensionGroup3D,
-    InventoryContextGroup3D,
-} from './InventoryContextGroup3D';
+import { MiniPortalContextGroup3D } from './MiniPortalContextGroup3D';
 import { CameraRig } from '../../shared/scene/CameraRigFactory';
 import { Game } from '../../shared/scene/Game';
 import { BotDimensionEvent } from '@casual-simulation/aux-vm';
 import { Color, Texture } from '@casual-simulation/three';
 import { DimensionGroup3D } from '../../shared/scene/DimensionGroup3D';
 import { PlayerSimulation3D } from './PlayerSimulation3D';
-import { InventoryPortalConfig } from './InventoryPortalConfig';
+import { MiniPortalConfig } from './MiniPortalConfig';
 
-export class InventorySimulation3D extends PlayerSimulation3D {
+export class MiniSimulation3D extends PlayerSimulation3D {
     /**
-     * The inventory dimension that this simulation is for.
+     * The mini portal dimension that this simulation is for.
      */
-    inventoryDimension: string;
+    miniDimension: string;
 
     getDefaultGridScale(): number {
-        return this.inventoryConfig.gridScale;
+        return this.miniConfig.gridScale;
     }
 
-    get inventoryConfig() {
-        return <InventoryPortalConfig>this.getPortalConfig('inventoryPortal');
+    get miniConfig() {
+        return <MiniPortalConfig>this.getPortalConfig(MINI_PORTAL);
     }
 
     get cameraControlsMode() {
-        return (
-            this.inventoryConfig.cameraControlsMode ?? super.cameraControlsMode
-        );
+        return this.miniConfig.cameraControlsMode ?? super.cameraControlsMode;
     }
 
     /**
      * Gets the background color that the simulation defines.
      */
     get backgroundColor() {
-        return this.inventoryConfig.backgroundColor || super.backgroundColor;
+        return this.miniConfig.backgroundColor || super.backgroundColor;
     }
 
     get backgroundAddress() {
-        return (
-            this.inventoryConfig.backgroundAddress || super.backgroundAddress
-        );
+        return this.miniConfig.backgroundAddress || super.backgroundAddress;
     }
 
     /**
-     * Gets the pannability of the inventory camera that the simulation defines.
+     * Gets the pannability of the mini camera that the simulation defines.
      */
     get pannable() {
-        return this.inventoryConfig.pannable;
+        return this.miniConfig.pannable;
     }
 
     /**
      * Gets the minimum value the pan can be set to on the x axis
      */
     get panMinX() {
-        return this.inventoryConfig.panMinX;
+        return this.miniConfig.panMinX;
     }
 
     /**
      * Gets the maximum value the pan can be set to on the x axis
      */
     get panMaxX() {
-        return this.inventoryConfig.panMaxX;
+        return this.miniConfig.panMaxX;
     }
 
     /**
      * Gets the minimum value the pan can be set to on the y axis
      */
     get panMinY() {
-        return this.inventoryConfig.panMinY;
+        return this.miniConfig.panMinY;
     }
 
     /**
      * Gets the maximum value the pan can be set to on the y axis
      */
     get panMaxY() {
-        return this.inventoryConfig.panMaxY;
+        return this.miniConfig.panMaxY;
     }
 
     /**
-     * Gets if rotation is allowed in the inventory that the simulation defines.
+     * Gets if rotation is allowed in the mini portal that the simulation defines.
      */
     get rotatable() {
-        return this.inventoryConfig.rotatable;
+        return this.miniConfig.rotatable;
     }
 
     /**
-     * Gets if zooming is allowed in the inventory that the simulation defines.
+     * Gets if zooming is allowed in the mini portal that the simulation defines.
      */
     get zoomable() {
-        return this.inventoryConfig.zoomable;
+        return this.miniConfig.zoomable;
     }
 
     /**
      * Gets the minimum value the zoom can be set to
      */
     get zoomMin() {
-        return this.inventoryConfig.zoomMin;
+        return this.miniConfig.zoomMin;
     }
 
     /**
      * Gets the maximum value the zoom can be set to
      */
     get zoomMax() {
-        return this.inventoryConfig.zoomMax;
+        return this.miniConfig.zoomMax;
     }
 
     /**
      * Gets the zoom level of the player that the simulation defines.
      */
     get playerZoom() {
-        return this.inventoryConfig.playerZoom;
+        return this.miniConfig.playerZoom;
     }
 
     /**
      * Gets the x-axis rotation of the player that the simulation defines.
      */
     get playerRotationX() {
-        return this.inventoryConfig.playerRotationX;
+        return this.miniConfig.playerRotationX;
     }
 
     /**
      * Gets the x-axis rotation of the player that the simulation defines.
      */
     get playerRotationY() {
-        return this.inventoryConfig.playerRotationY;
+        return this.miniConfig.playerRotationY;
     }
 
     /**
      * Gets whether the portal is resizable.
      */
     get resizable() {
-        return this.inventoryConfig.resizable;
+        return this.miniConfig.resizable;
     }
 
     /**
      * Gets the height of the portal.
      */
     get height() {
-        return this.inventoryConfig.height;
+        return this.miniConfig.height;
+    }
+
+    /**
+     * Gets the width of the portal.
+     */
+    get width() {
+        return this.miniConfig.width;
     }
 
     /**
      * Gets whether to show the camera focus point.
      */
     get showFocusPoint() {
-        return this.inventoryConfig.showFocusPoint;
+        return this.miniConfig.showFocusPoint;
+    }
+
+    /**
+     * Gets the style the cursor should have for this portal.
+     */
+    get cursor() {
+        return this.miniConfig.cursor;
     }
 
     constructor(game: Game, simulation: BrowserSimulation) {
-        super('inventoryPortal', game, simulation);
+        super(MINI_PORTAL, game, simulation);
     }
 
     getMainCameraRig(): CameraRig {
-        return this._game.getInventoryCameraRig();
+        return this._game.getMiniPortalCameraRig();
     }
 
     init() {
@@ -186,14 +194,13 @@ export class InventorySimulation3D extends PlayerSimulation3D {
                 .pipe(
                     filter((bot) => !!bot),
                     tap((bot) => {
-                        const userInventoryDimensionValue =
-                            bot.values['inventoryPortal'];
-                        const previousDimension = this.inventoryDimension;
-                        this.inventoryDimension = userInventoryDimensionValue;
-                        if (previousDimension !== userInventoryDimensionValue) {
+                        const userMiniDimensionValue = bot.values[MINI_PORTAL];
+                        const previousDimension = this.miniDimension;
+                        this.miniDimension = userMiniDimensionValue;
+                        if (previousDimension !== userMiniDimensionValue) {
                             console.log(
-                                '[InventorySimulation3D] User changed inventory dimension to: ',
-                                userInventoryDimensionValue
+                                '[MiniSimulation3D] User changed mini portal dimension to: ',
+                                userMiniDimensionValue
                             );
                         }
                     })
@@ -204,7 +211,7 @@ export class InventorySimulation3D extends PlayerSimulation3D {
     }
 
     protected _constructDimensionGroup(portalTag: string, bot: Bot) {
-        return new InventoryContextGroup3D(
+        return new MiniPortalContextGroup3D(
             this,
             bot,
             'player',
@@ -214,8 +221,8 @@ export class InventorySimulation3D extends PlayerSimulation3D {
     }
 
     protected _createPortalConfig(portalTag: string) {
-        if (portalTag === 'inventoryPortal') {
-            return new InventoryPortalConfig(portalTag, this.simulation);
+        if (portalTag === MINI_PORTAL) {
+            return new MiniPortalConfig(portalTag, this.simulation);
         } else {
             return super._createPortalConfig(portalTag);
         }
