@@ -12,6 +12,8 @@ import {
     BotCalculationContext,
     Bot,
     BotTags,
+    getBotCursor,
+    hasValue,
 } from '@casual-simulation/aux-common';
 import { Physics } from '../scene/Physics';
 import { flatMap } from 'lodash';
@@ -176,6 +178,20 @@ export abstract class BaseInteractionManager {
         });
     }
 
+    /**
+     * Removes and disposes of any operations that the given filter returns true for.
+     * @param filter
+     */
+    clearOperations(filter: (operation: IOperation) => boolean) {
+        this._operations = this._operations.filter((o) => {
+            if (filter(o)) {
+                o.dispose();
+                return false;
+            }
+            return true;
+        });
+    }
+
     update(): void {
         // const calc = appManager.simulationManager.primary.helper.createContext();
         // Update active operations and dispose of any that are finished.
@@ -229,9 +245,6 @@ export abstract class BaseInteractionManager {
             this.setCameraControlsEnabled(this._cameraControlsEnabled);
         }
 
-        this._updateCameraOffsets();
-        this._updateCameraControls();
-
         // Detect left click.
         this._handleMouseInput(input);
         this._handleControllerInput(input);
@@ -243,11 +256,34 @@ export abstract class BaseInteractionManager {
         this._updateHoveredBots();
         this._updateFocusedBots();
         this._updatePlayerBotTags();
+        this._updateCursors();
+
+        // Update camera controls after handing other inputs so that
+        // we can enable/disable the controls before it tries using the input.
+        this._updateCameraOffsets();
+        this._updateCameraControls();
     }
 
     protected _updatePlayerBotTags() {}
 
     protected _updateCameraOffsets() {}
+
+    protected _updateCursors() {
+        let hasCursor = false;
+        for (let bot of this._hoveredBots) {
+            const cursor = getBotCursor(null, bot.bot);
+            if (hasValue(cursor)) {
+                // TODO: rework to do a better job of handling
+                this._game.botCursor = cursor;
+                hasCursor = true;
+                break;
+            }
+        }
+
+        if (!hasCursor) {
+            this._game.botCursor = null;
+        }
+    }
 
     protected _updateCameraControls() {
         for (let controller of this._cameraRigControllers) {
