@@ -8,6 +8,7 @@ import { Game } from '../scene/Game';
 import { SubscriptionLike } from 'rxjs';
 import { EventBus } from '../EventBus';
 import { debounce } from 'lodash';
+import { BotCursorType, getCursorCSS } from '@casual-simulation/aux-common';
 
 export interface SidebarItem {
     id: string;
@@ -26,6 +27,8 @@ export interface SidebarItem {
 export default class BaseGameView extends Vue implements IGameView {
     private _resizeObserver: import('@juggle/resize-observer').ResizeObserver;
     protected _subscriptions: SubscriptionLike[] = [];
+
+    cursor: string = null;
 
     _game: Game = null;
 
@@ -46,11 +49,12 @@ export default class BaseGameView extends Vue implements IGameView {
 
     async created() {
         this._subscriptions = [];
-        this.resize = debounce(this.resize.bind(this), 100);
+        const resize = this.resize.bind(this);
+        this.resize = debounce(resize, 100);
 
-        EventBus.$on('resize', this.resize);
+        EventBus.$on('resize', () => this.resize());
         // window.addEventListener('resize', this.resize);
-        window.addEventListener('vrdisplaypresentchange', this.resize);
+        window.addEventListener('vrdisplaypresentchange', () => this.resize());
 
         this._game = this.createGame();
 
@@ -87,6 +91,10 @@ export default class BaseGameView extends Vue implements IGameView {
         this._game = this.createGame();
         this._game.setup();
         this.resize();
+    }
+
+    setCursor(cursor: BotCursorType): void {
+        this.cursor = getCursorCSS(cursor);
     }
 
     beforeDestroy() {
@@ -142,10 +150,13 @@ export default class BaseGameView extends Vue implements IGameView {
         this._setWidthAndHeight(width, height);
     }
 
+    protected setWidthAndHeightCore(width: number, height: number) {}
+
     private _setWidthAndHeight(width: number, height: number) {
         this._game.onWindowResize(width, height);
 
         this.container.style.height = this.gameView.style.height = this._game.getRenderer().domElement.style.height;
         this.container.style.width = this.gameView.style.width = this._game.getRenderer().domElement.style.width;
+        this.setWidthAndHeightCore(width, height);
     }
 }
